#494 방장탈퇴못하게함

+++ src/components/my-clubs/info/ResignModal.tsx
@@ -5,6 +5,7 @@ import WithdrawalPopup from './WithdrawalPopup';
 import { MEMBERS_TABLE } from '@/common/constants/tableNames';
 import { createClient } from '@/utils/supabase/client';
 import { useRouter } from 'next/navigation';
+import ToastUi from '@/common/ToastUi';
 const ResignModal = ({
   isModal,
   onClose,
@@ -16,13 +17,33 @@ const ResignModal = ({
 }) => {
   const supabase = createClient();
   const router = useRouter();
-
+  const toastStyle = {
+    width: '343px',
+    height: '50px',
+    top: '200px',
+    left: '50%',
+    transform: 'translateX(-50%)',
+    fontSize: '8px'
+  };
+  const [isAdmin, setIsAdmin] = useState(false);
   const handleResign = async () => {
     const {
       data: { user }
     } = await supabase.auth.getUser();
     if (user) {
       try {
+        const { data } = await supabase
+          .from('members')
+          .select('role')
+          .eq('club_id', clubId)
+          .eq('user_id', user.id)
+          .single();
+        if (data?.role === 'admin') {
+          // alert('너 방장');
+          setIsAdmin(true);
+
+          return;
+        }
         const { error } = await supabase
           .from(MEMBERS_TABLE)
           .delete()
@@ -67,6 +88,17 @@ const ResignModal = ({
           isOpen={isOpenPopup}
           onClose={() => setIsOpenPopup(false)}
           onWithdraw={() => handleResign()}
+        />
+      )}
+      {isAdmin && (
+        <ToastUi
+          message='방장은 탈퇴할 수 없습니다.'
+          onClose={() => {
+            setIsAdmin(false);
+          }}
+          isSuccess={false}
+          duration={1000}
+          style={toastStyle}
         />
       )}
     </>
